+++ test_all.py
@@ -9,18 +9,14 @@ import unittest
 from test_basics import *
 from test_ashfall import *
 from test_cmdexecution import *
-from test_composition import *
-from test_exposure import *
 from test_fragility import *
 from test_gpdexposure import *
 from test_intensity import *
 from test_intensitydatawrapper import *
 from test_loss import *
 from test_performance import *
-from test_performance_optimization import *
 from test_schemamapping import *
 from test_shakemap import *
-from test_transition import *
 
 
 if __name__ == "__main__":