+++ test_gpdexposure.py
@@ -39,7 +39,9 @@ class TestGpdExposureDamageStateUpdate(unittest.TestCase):
         expo = {
             'Taxonomy': ['TAX1', 'TAX1', 'TAX2', 'TAX2'],
             'Damage': ['D0', 'D1', 'D0', 'D1'],
-            'Buildings': [100.0, 100.0, 100.0, 100.0]
+            'Buildings': [100.0, 100.0, 100.0, 100.0],
+            'Population': [20.0, 10.0, 20.0, 10.0],
+            'Repl-cost-USD-bdg': [50000, 45000, 60000, 59000]
         }
 
         series = pandas.Series({
@@ -173,22 +175,26 @@ class TestGpdExposureDamageStateUpdate(unittest.TestCase):
                 'target_taxonomy': 'TAX',
                 'conv_matrix': {
                     '0': {
+                        # TAX1 D0 goes to 100% into TAX D0
                         '0': 1.0,
                         '1': 0.0,
                         '2': 0.0,
                     },
                     '1': {
                         '0': 0.0,
+                        # TAX1 D1 goes to 50% into TAX D1
                         '1': 0.5,
                         '2': 0.2,
                     },
                     '2': {
                         '0': 0.0,
+                        # TAX1 D1 goes to 30% into TAX D2
                         '1': 0.3,
                         '2': 0.3,
                     },
                     '3': {
                         '0': 0.0,
+                        # TAX1 D1 goes to 20% into TAX D3
                         '1': 0.2,
                         '2': 0.5,
                     }
@@ -253,10 +259,132 @@ class TestGpdExposureDamageStateUpdate(unittest.TestCase):
         self.assertEqual(4, len(expo))
 
         # This here was the mapping only
+        # (but with applying the damage computation)
         self.assertEqual(24, get_buildings(expo, 'TAX', 'D0'))
         self.assertBetween(139.19, get_buildings(expo, 'TAX', 'D1'), 140.01)
         self.assertBetween(140.79, get_buildings(expo, 'TAX', 'D2'), 140.81)
         self.assertEqual(96, get_buildings(expo, 'TAX', 'D3'))
+
+        # Population follows the same algorithm as the number of buildings
+        self.assertBetween(4.79, get_population(expo, 'TAX', 'D0'), 4.81)
+        self.assertBetween(21.11, get_population(expo, 'TAX', 'D1'), 21.13)
+        self.assertBetween(20.47, get_population(expo, 'TAX', 'D2'), 20.49)
+        self.assertBetween(13.5, get_population(expo, 'TAX', 'D3'), 13.7)
+
+        # The replacement costs don't follow the route of building handling.
+        #
+        # First we calculate the total replacement costs per class
+        #
+        # totalRepl('TAX1', 'D0') = replBdg('TAX1', 'D0') * nBdg('TAX1', 'D0')
+        #                         = 50_000 * 100
+        #                         = 5_000_000
+        # totalRepl('TAX1', 'D1') = 4_500_000
+        # totalRepl('TAX2', 'D0') = 6_000_000
+        # totalRepl('TAX2', 'D1') = 5_900_000
+        #
+        # Then we map the total replacement costs to the other schemas:
+        #
+        # totalRepl('TAX', 'D0') = (
+        #                            totalRepl('TAX1', 'D0') *
+        #                              p('TAX1', 'D0', 'TAX', 'D0') +
+        #                            totalRepl('TAX2', 'D0') *
+        #                              p('TAX2', 'D0', 'TAX', 'D0')
+        #                          )
+        #                        = 5_000_000 * 1 + 6_000_000 * 1
+        # totalRepl('TAX', 'D0') = 11_000_000
+        # totalRepl('TAX', 'D1') = (
+        #                            totalRepl('TAX1', 'D1') *
+        #                              p('TAX1', 'D1', 'TAX', 'D1') +
+        #                            totalRepl('TAX2', 'D1') *
+        #                              p('TAX2', 'D1', 'TAX', 'D1')
+        #                          )
+        #                        = 4_500_000 * 0.5 + 5_900_000 * 0.9
+        #                        = 7_560_000
+        # totalRepl('TAX', 'D2') = (
+        #                            totalRepl('TAX1', 'D1') *
+        #                              p('TAX1', 'D1', 'TAX', 'D2') +
+        #                            totalRepl('TAX2', 'D1') *
+        #                              p('TAX2', 'D1', 'TAX', 'D2')
+        #                          )
+        #                        = 4_500_000 * 0.3 + 5_900_000 * 0.1
+        #                        = 1_940_000
+        # totalRepl('TAX', 'D3') = totalRepl('TAX1', 'D1') *
+        #                          p('TAX1', 'D1', 'TAX', 'D3')
+        #                        = 4_500_000 * 0.2
+        #                        = 900_000
+        #
+        # Then we sum those up.
+        #
+        # totalRepl('TAX') = (
+        #                      totalRepl('TAX', 'D0') +
+        #                      totalRepl('TAX', 'D1') +
+        #                      totalRepl('TAX', 'D2') +
+        #                      totalRepl('TAX', 'D3')
+        #                    )
+        #                  = 21_400_000
+        #
+        # And last we calculate the replacement cost per building
+        #
+        # replBdg('TAX') = totalRepl('TAX') / nBdg('TAX')
+        # replBdg('TAX') = 21_400_000 / 400
+        # replBdg('TAX') = 53_500
+        #
+        # As we are not going to change the replacement costs on applying
+        # the damage, we stay with those values.
+        #
+        self.assertBetween(
+            53_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX', 'D0'),
+            53_501
+        )
+        self.assertBetween(
+            53_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX', 'D1'),
+            53_501
+        )
+        self.assertBetween(
+            53_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX', 'D2'),
+            53_501
+        )
+        self.assertBetween(
+            53_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX', 'D3'),
+            53_501
+        )
+
+        transitions = pandas.DataFrame(result_exposure.iloc[0].transitions)
+
+        self.assertBetween(
+            71,
+            get_transition_n_bdg(transitions, 'TAX', 0, 1),
+            73
+        )
+        self.assertBetween(
+            63,
+            get_transition_n_bdg(transitions, 'TAX', 0, 2),
+            65
+        )
+        self.assertBetween(
+            39,
+            get_transition_n_bdg(transitions, 'TAX', 0, 3),
+            41
+        )
+        self.assertBetween(
+            44.7,
+            get_transition_n_bdg(transitions, 'TAX', 1, 2),
+            44.9
+        )
+        self.assertBetween(
+            27.9,
+            get_transition_n_bdg(transitions, 'TAX', 1, 3),
+            28.1
+        )
+        self.assertBetween(
+            7.9,
+            get_transition_n_bdg(transitions, 'TAX', 2, 3),
+            8.1
+        )
 
     def test_without_schema_mapping(self):
         """
@@ -283,6 +411,89 @@ class TestGpdExposureDamageStateUpdate(unittest.TestCase):
         self.assertBetween(104.99, get_buildings(expo, 'TAX2', 'D1'), 105.01)
         self.assertBetween(79.99, get_buildings(expo, 'TAX2', 'D2'), 80.01)
 
+        # Normally the replacement costs per building would stay the very same
+        # as the input values (as all of those are only taxonomy specific and
+        # oriented on the replacement costs of D0).
+        # However, as there is the danger, that the replacement costs are
+        # different over the damage states (having not completely reliable
+        # input), we are going to compute weighted means before applying
+        # the damage state conversion.
+        #
+        # Again we go over the total replacement costs:
+        # totalRepl('TAX1', 'D0') = replBdg('TAX1', 'D0') * nBdg('TAX1', 'D0')
+        # ...
+        # Then we sum them up
+        # totalRepl('TAX1') = totalRepl('TAX1', 'D0') + totalRepl('TAX1', 'D1')
+        #
+        # And then we calculate the replacement costs per building
+        # replBdg('TAX1') = totalRepl('TAX1') / nBdg('TAX1')
+        # And those are the very same regardless of the damage state
+
+        self.assertBetween(
+            47_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX1', 'D0'),
+            47_501
+        )
+        self.assertBetween(
+            47_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX1', 'D1'),
+            47_501
+        )
+        self.assertBetween(
+            47_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX1', 'D2'),
+            47_501
+        )
+
+        self.assertBetween(
+            59_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX2', 'D0'),
+            59_501
+        )
+        self.assertBetween(
+            59_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX2', 'D1'),
+            59_501
+        )
+        self.assertBetween(
+            59_499,
+            get_replacement_costs_usd_bdg(expo, 'TAX2', 'D2'),
+            59_501
+        )
+
+        transitions = pandas.DataFrame(result_exposure.iloc[0].transitions)
+
+        self.assertBetween(
+            37.49,
+            get_transition_n_bdg(transitions, 'TAX1', 0, 1),
+            37.51
+        )
+        self.assertBetween(
+            24,
+            get_transition_n_bdg(transitions, 'TAX1', 0, 2),
+            26
+        )
+        self.assertBetween(
+            24,
+            get_transition_n_bdg(transitions, 'TAX1', 1, 2),
+            26
+        )
+        self.assertBetween(
+            44.9,
+            get_transition_n_bdg(transitions, 'TAX2', 0, 1),
+            45.1
+        )
+        self.assertBetween(
+            39.9,
+            get_transition_n_bdg(transitions, 'TAX2', 0, 2),
+            40.1
+        )
+        self.assertBetween(
+            39.9,
+            get_transition_n_bdg(transitions, 'TAX2', 1, 2),
+            40.1
+        )
+
     def assertBetween(self, lower, x, upper):
         """
         Test that a number is between two others.
@@ -298,6 +509,29 @@ def get_buildings(expo, tax, ds):
     expo_tax = expo[expo.Taxonomy == tax]
     expo_ds = expo_tax[expo_tax.Damage == ds]
     return expo_ds.Buildings.iloc[0]
+
+
+def get_population(expo, tax, ds):
+    """Get the population for the tax and the ds."""
+    expo_tax = expo[expo.Taxonomy == tax]
+    expo_ds = expo_tax[expo_tax.Damage == ds]
+    return expo_ds.Population.iloc[0]
+
+
+def get_replacement_costs_usd_bdg(expo, tax, ds):
+    """Get the replacement costs in usd per bdg."""
+    expo_tax = expo[expo.Taxonomy == tax]
+    expo_ds = expo_tax[expo_tax.Damage == ds]
+    return expo_ds['Repl-cost-USD-bdg'].iloc[0]
+
+
+def get_transition_n_bdg(transitions, tax, from_ds, to_ds):
+    """Return the number of buildings in the transition."""
+    filter_1 = transitions[transitions.taxonomy == tax]
+    filter_2 = filter_1[filter_1.from_damage_state == from_ds]
+    filter_3 = filter_2[filter_2.to_damage_state == to_ds]
+
+    return filter_3.n_buildings.iloc[0]
 
 
 class MakeFakeFunction: